調整角色視窗樣式與欄位型別，優化使用體驗
- 將 Modal 樣式調整為 90% viewport 並支援內容捲動，提升大視窗下的操作體驗
- RecordVm 的 PainScore 型別由 int 改為 double，支援小數分數
- Modal 標題改用變數綁定，內容外層加上專用 class 以利樣式調整
- 例外處理日誌層級由 Error 降為 Warning

+++ src/MyProject/MyProject.Models/Others/RecordVm.cs
@@ -20,7 +20,7 @@ public class RecordVm
     public string Department { get; set; } = "OPD";
 
     [Range(0, 10)]
-    public int PainScore { get; set; } = 0;     // 拉霸常用 0~10
+    public double PainScore { get; set; } = 0;     // 拉霸常用 0~10
 
     [StringLength(500)]
     public string Notes { get; set; } = "";
